src/requirements/base.txt
# Core
Django>=1.8,<1.9
Pillow>=3.1,<3.2
psycopg2>=2.6,<2.7

# Packages
apns>=2.0,<2.1
celery[librabbitmq]>=3.1,<3.2
django-admin-tools>=0.7,<0.8
django-cors-headers>=1.1,<1.2
django-extensions>=1.6,<1.7
django-filter>=0.11,<0.12
django-rest-swagger>=0.3,<0.4
djangorestframework-jwt>=1.7,<1.8
djangorestframework>=3.2,<3.3
drf-nested-routers>=0.10,<0.11
Markdown>=2.6,<2.7
stripe>=1.28,<1.29
Werkzeug>=0.11,<0.12

# Reports
pyexcel
pyexcel-xlsx

# django channels
channels
asgi_redis

# Tools
ipython
unicodecsv
